Add assertions
Added assertions into the Storage class to ensure that
the storageFile is not null, and parsedData length is correct.

Added assertions into the Ui class to ensure that the tasks
found with the keyword is not null.

Added assertions into the MainWindow class to ensure that
the provided chatbot Jimmy is not null.